+++ docs/changes.md
@@ -8,6 +8,7 @@ title: Tenanti Change Log
 ### v2.2.1@dev {#v2-2-1}
 
 * Allow to specifically declare database connection name when running `Orchestra\Tenanti\Observer`. This allow user to setup multi-tenancy on a different database.
+* Allow to configuration custom multi-tenant migration table name.
 
 ### v2.2.0 {#v2-2-0}