+++ docs/changes.md
@@ -5,7 +5,7 @@ title: Tenanti Change Log
 
 ## Version 2.2 {#v2-2}
 
-### v2.2.1@dev {#v2-2-1}
+### v2.2.1 {#v2-2-1}
 
 * Allow to specifically declare database connection name when running `Orchestra\Tenanti\Observer`. This allow user to setup multi-tenancy on a different database.
 * Allow to configuration custom multi-tenant migration table name.